refactor: move test to correct spot
